requirements.txt
asgiref==3.11.1
dj-database-url==3.1.2
Django==6.0.3
mysqlclient==2.2.8
pillow==12.1.1
psycopg2-binary==2.9.11
python-dotenv==1.2.2
sqlparse==0.5.5
tzdata==2025.3
